Separated invocation builder from specific implementations.
commit-id:a670d70f
